Use BuildKit and branching Dockerfile
Locally it would be better to be able to build the project internally in
the Dockerfile so it cooperates better with tools like Docker-Compose.

But as long as Jenkins is our CI, it is a hassle setting up proxy and
certificates inside the docker build process, so there it is better to
build on the outside, and copy built artifacts in.

BuildKit allows branched Dockerfiles, and will only build the necessary
branches. This PR makes a separate branch for building inside Docker, so
docker-compose can work seamlessly.